% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.check_ccpy_installation}
\alias{.check_ccpy_installation}
\title{Check if CloudComPy is properly installed.}
\usage{
.check_ccpy_installation()
}
\value{
Logical indicating if CloudComPy installation is valid
}
\description{
Check if CloudComPy is properly installed.
}
\keyword{internal}
