% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cloudcompy.R
\name{install_cloudcompy}
\alias{install_cloudcompy}
\title{Install CloudComPy from a bin}
\usage{
install_cloudcompy(cloudcompy)
}
\arguments{
\item{cloudcompy}{The path to the CloudComPy archive}
}
\description{
Unpacks the CloudComPy bin into your user directory, creates a miniconda
instance there and creates a conda environment for CloudComPy.
Currently only a single version of ClouComPy is supported (20240613).
Download it from here:
\url{https://www.simulation.openfields.fr/index.php/cloudcompy-downloads/
3-cloudcompy-binaries/5-windows-cloudcompy-binaries/
106-cloudcompy310-20240613>}
}
